move image build and container run functions into lower namespace; remove docker as a dependency
